requirements.txt

Flask==2.1.1
Werkzeug==2.1.1
pymongo==3.12.3
gridfs==1.0.1
pydub==0.25.1
SpeechRecognition==3.8.1
python-dotenv==0.20.0
